pyproject.toml
[tool.poetry]
name = "syn"
version = "0.1.0"
description = "Code for experiments in \"Boosting Data Analytics with Synthetic Volume Expansion\" by X.Shen, Y. Liu and R. Shen."
authors = ["Yifei Liu <liu00980@umn.edu>", "Rex Shen <rshen0@stanford.edu>", "Xiaotong Shen <xshen@umn.edu >"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
torch = {url = "https://download.pytorch.org/whl/cu102/torch-1.12.1%2Bcu102-cp39-cp39-linux_x86_64.whl"}
torchaudio = {url = "https://download.pytorch.org/whl/cu102/torchaudio-0.12.1%2Bcu102-cp39-cp39-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu102/torchvision-0.13.1%2Bcu102-cp39-cp39-linux_x86_64.whl"}
diffusers = "0.15.0"
torchmetrics = {extras = ["image"], version = "^1.2.0"}
transformers = "4.27.4"
pytorch-lightning = "^2.1.1"
tensorboard = "^2.15.1"
catboost = "1.0.3"
category-encoders = "2.3.0"
dython = "0.5.1"
icecream = "2.1.2"
libzero = "0.0.8"
optuna = "2.10.1"
pandas = "1.3.4"
pyarrow = "6.0.0"
rtdl = "0.0.9"
scikit-learn = "1.0.2"
skorch = "0.11.0"
tomli-w = "0.4.0"
tomli = "1.2.2"
tqdm = "4.62.3"
scipy = "1.8.0"
setuptools = "^69.1.1"
gdown = "^5.1.0"
sqlalchemy = "1.4.46"
python-dotenv = "^1.0.1"
pot = "0.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
